fix(polls): rename PollEditor
